Connect RN reset pin to PA4
The NRST pin is pulled up using the pull-up resistor R1. It's also
connected to the PA4 pin of the microcontroller.

By default, after reset, the GPIO pins of the uC are configured as
floating inputs (see datasheet table 14). If we enable the pull-down
resistor on PA4, we should be able to reset the RN module.

Refs #13